+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ## Added
+- extended list of supported formats
 - support docxf and oform formats
 - support for the connector to work in conjunction with the plugin "Force Lowercase URLs" (https://wordpress.org/plugins/force-lowercase-urls)
 - setting authorization header